[SLP][modularisation][NFC] Move isFixedVectorShuffle to SLPUtils
Move the BoUpSLP-independent shuffle helper out of SLPVectorizer.cpp
into SLPVectorizer/SLPUtils.{h,cpp}:

  isFixedVectorShuffle

It sits with the existing shuffle/mask helpers. Behavior is unchanged.

Part of the SLPVectorizer.cpp modularization effort:
https://discourse.llvm.org/t/modularizing-slpvectorizer-cpp/90922
